refactor cv display comps
