+++ app/app/cvs/_components/DocList.tsx
@@ -1,33 +1,13 @@
 "use client";
 
-import { useState } from "react";
-
-import {
-  useDocContentQuery,
-  useUserDocsQuery,
-} from "@/actions/queries/user.queries";
+import { useUserDocsQuery } from "@/actions/queries/user.queries";
 import { formatDistanceToNow } from "date-fns";
-import { MoreHorizontal } from "lucide-react";
-import { toast } from "react-hot-toast";
 
 import {
   ErrorFallback,
   NoDataFallback,
 } from "@/components/global/QueryFallbacks";
 import TruncatedText from "@/components/global/TuncateText";
-import { Button } from "@/components/ui/button";
-import {
-  Dialog,
-  DialogContent,
-  DialogHeader,
-  DialogTitle,
-} from "@/components/ui/dialog";
-import {
-  DropdownMenu,
-  DropdownMenuContent,
-  DropdownMenuItem,
-  DropdownMenuTrigger,
-} from "@/components/ui/dropdown-menu";
 import {
   Table,
   TableBody,
@@ -37,8 +17,7 @@ import {
   TableRow,
 } from "@/components/ui/table";
 
-import { MdToPdf } from "@/lib/utils";
-
+import DocContextMenu from "./ContextMenu";
 import Skeleton from "./Skeleton";
 
 interface GeneratedDocsListProps {
@@ -93,84 +72,6 @@ const GeneratedDocsList = ({ userId }: GeneratedDocsListProps) => {
         </TableBody>
       </Table>
     </div>
-  );
-};
-
-interface DocContextMenuProps {
-  docId: string;
-  notes: string;
-}
-
-const DocContextMenu = ({ docId, notes }: DocContextMenuProps) => {
-  const [isNotesOpen, setIsNotesOpen] = useState(false);
-  const [isDownloading, setIsDownloading] = useState(false);
-
-  const { refetch: fetchContent } = useDocContentQuery(docId, false);
-
-  const handleDownload = async () => {
-    setIsDownloading(true);
-    const loadingToast = toast.loading("Fetching document...");
-
-    try {
-      const result = await fetchContent();
-      if (result.data?.success) {
-        console.log("Document content:", result.data.docContent);
-        toast.success("Document fetched successfully");
-      } else {
-        toast.error("Failed to fetch document");
-      }
-
-      if (
-        result.data?.docContent?.content &&
-        result.data?.docContent?.documentTitle
-      ) {
-        MdToPdf(
-          result.data.docContent.content,
-          result.data.docContent.documentTitle,
-        );
-      } else {
-        toast.error("Invalid document data");
-      }
-    } catch (error) {
-      console.error("Download error:", error);
-      toast.error("Error downloading document");
-    } finally {
-      toast.dismiss(loadingToast);
-      setIsDownloading(false);
-    }
-  };
-
-  return (
-    <>
-      <DropdownMenu>
-        <DropdownMenuTrigger asChild>
-          <Button variant="ghost" className="h-8 w-8 p-0">
-            <span className="sr-only">Open menu</span>
-            <MoreHorizontal className="h-4 w-4" />
-          </Button>
-        </DropdownMenuTrigger>
-        <DropdownMenuContent align="end">
-          <DropdownMenuItem onSelect={() => setIsNotesOpen(true)}>
-            Notes
-          </DropdownMenuItem>
-          <DropdownMenuItem onClick={handleDownload} disabled={isDownloading}>
-            Download
-          </DropdownMenuItem>
-        </DropdownMenuContent>
-      </DropdownMenu>
-
-      <Dialog open={isNotesOpen} onOpenChange={setIsNotesOpen}>
-        <DialogContent className="p-6 sm:max-w-[700px]">
-          <DialogHeader>
-            <DialogTitle className="text-xl">Notes</DialogTitle>
-          </DialogHeader>
-
-          <div className="max-h-[60vh] overflow-y-auto whitespace-pre-wrap pr-2 text-sm leading-relaxed text-muted-foreground">
-            {notes}
-          </div>
-        </DialogContent>
-      </Dialog>
-    </>
   );
 };
