+++ app/app/cvs/_components/Skeleton.tsx
@@ -9,9 +9,25 @@ import {
   TableRow,
 } from "@/components/ui/table";
 
-function Skeleton() {
-  return (
-    <div className="bg-background/30 border rounded-md">
+function Skeleton({ isAdmin = false }: { isAdmin?: boolean }) {
+  return isAdmin ? (
+    <div className="rounded-md border bg-background/30">
+      <Table>
+        <TableHeader>
+          <TableRow>
+            <TableHead>User</TableHead>
+            <TableHead>User Email</TableHead>
+            <TableHead>Candidate</TableHead>
+            <TableHead>Location</TableHead>
+            <TableHead>Created</TableHead>
+            <TableHead className="w-[50px]"></TableHead>
+          </TableRow>
+        </TableHeader>
+        <TableBody>{renderTableRows(isAdmin)}</TableBody>
+      </Table>
+    </div>
+  ) : (
+    <div className="rounded-md border bg-background/30">
       <Table>
         <TableHeader>
           <TableRow>
@@ -23,33 +39,46 @@ function Skeleton() {
             <TableHead className="w-[50px]"></TableHead>
           </TableRow>
         </TableHeader>
-        <TableBody>
-          {Array.from({ length: 10 }).map((_, index) => (
-            <TableRow key={index}>
-              <TableCell>
-                <div className="bg-neutral-500 rounded w-32 h-2 animate-pulse"></div>
-              </TableCell>
-              <TableCell>
-                <div className="bg-neutral-500 rounded w-32 h-2 animate-pulse"></div>
-              </TableCell>
-              <TableCell>
-                <div className="bg-neutral-500 rounded w-20 h-2 animate-pulse"></div>
-              </TableCell>
-              <TableCell>
-                <div className="bg-neutral-500 rounded w-24 h-2 animate-pulse"></div>
-              </TableCell>
-              <TableCell>
-                <div className="bg-neutral-500 rounded w-32 h-2 animate-pulse"></div>
-              </TableCell>
-              <TableCell>
-                <div className="bg-neutral-500 rounded w-8 h-2 animate-pulse"></div>
-              </TableCell>
-            </TableRow>
-          ))}
-        </TableBody>
+        <TableBody>{renderTableRows(isAdmin)}</TableBody>
       </Table>
     </div>
   );
 }
+
+const renderTableRows = (isAdmin: boolean) =>
+  Array.from({ length: 10 }).map((_, index) => (
+    <TableRow key={index}>
+      {isAdmin ? (
+        <TableCell className="font-medium">
+          <div className="flex items-center gap-3">
+            <div className="h-8 w-8 animate-pulse rounded-full bg-neutral-500"></div>
+            <div className="flex flex-col gap-2">
+              <div className="h-2 w-32 animate-pulse rounded bg-neutral-500"></div>
+              <div className="h-2 w-24 animate-pulse rounded bg-neutral-500"></div>
+            </div>
+          </div>
+        </TableCell>
+      ) : (
+        <TableCell>
+          <div className="h-2 w-32 animate-pulse rounded bg-neutral-500"></div>
+        </TableCell>
+      )}
+      <TableCell>
+        <div className="h-2 w-32 animate-pulse rounded bg-neutral-500"></div>
+      </TableCell>
+      <TableCell>
+        <div className="h-2 w-32 animate-pulse rounded bg-neutral-500"></div>
+      </TableCell>
+      <TableCell>
+        <div className="h-2 w-32 animate-pulse rounded bg-neutral-500"></div>
+      </TableCell>
+      <TableCell>
+        <div className="h-2 w-20 animate-pulse rounded bg-neutral-500"></div>
+      </TableCell>
+      <TableCell>
+        <div className="h-2 w-8 animate-pulse rounded bg-neutral-500"></div>
+      </TableCell>
+    </TableRow>
+  ));
 
 export default Skeleton;
